[proc] split cgroup linux test to separate file to avoid test fails
