+++ pkg/util/docker/cgroup.go
@@ -6,8 +6,6 @@
 // This code is not tied to docker itself, hence no docker build flag.
 // It could be moved to its own package.
 
-// +build docker
-
 package docker
 
 import (
