+++ pkg/util/docker/cgroup_test.go
@@ -3,8 +3,6 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2017 Datadog, Inc.
 
-// +build docker
-
 package docker
 
 import (
@@ -98,19 +96,6 @@ func TestMemLimit(t *testing.T) {
 	tempFolder.add("memory/memory.limit_in_bytes", "1234")
 	value, err = cgroup.MemLimit()
 	assert.Nil(t, err)
-	assert.Equal(t, value, uint64(1234))
-}
-
-// TestDindContainer is to test if our agent can handle dind container correctly
-func TestDindContainer(t *testing.T) {
-	containerID := "6ab998413f7ae63bb26403dfe9e7ec02aa92b5cfc019de79da925594786c985f"
-	tempFolder, cgroup, err := newDindContainerCgroup("dind-container", "memory", containerID)
-	assert.NoError(t, err)
-	tempFolder.add("memory.limit_in_bytes", "1234")
-	defer tempFolder.removeAll()
-
-	value, err := cgroup.MemLimit()
-	assert.NoError(t, err)
 	assert.Equal(t, value, uint64(1234))
 }
